cast turtle x,y to int

+++ TurtleArt/tacanvas.py
@@ -571,8 +571,8 @@ class TurtleGraphics:
         x, y = self.width / 2 + int(self.xcor), \
                self.height / 2 - int(self.ycor)
         self.tw.active_turtle.move(
-            (self.cx + x - self.tw.active_turtle.spr.rect.width/2,
-             self.cy + y - self.tw.active_turtle.spr.rect.height/2))
+            (int(self.cx + x - self.tw.active_turtle.spr.rect.width/2),
+             int(self.cy + y - self.tw.active_turtle.spr.rect.height/2)))
 
     def invalt(self, x, y, w, h):
         """ Mark a region for refresh """
